server/scripts/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
psycopg[binary]>=3.1.0
lxml>=4.9.0
html5lib>=1.1 